import optimization, bug fix
