+++ src/components/Form/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useConfigForm } from "./useConfigForm";
 export { useConfigData } from "./useConfigData";
+export type { TMockData } from "./types";
