+++ src/components/Form/hooks/useConfigForm.ts
@@ -1,8 +1,8 @@
 import { SubmitHandler, useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 
-import { schema } from "../validation";
 import { transformData, convertingValues, updateValues } from "../utils";
+import { schema } from "@components";
 import type { Inputs } from "@components";
 import type { TBackendData } from "../data";
 
@@ -17,13 +17,13 @@ export const useConfigForm = (data: TBackendData) => {
     control,
     setValue,
     formState: { errors, isDirty },
-  } = useForm<Inputs>({
+  } = useForm({
     mode: "all",
     defaultValues,
     resolver: zodResolver(schema),
   });
 
-  const onSubmit: SubmitHandler<Inputs> = (data) => {
+  const onSubmit: SubmitHandler<Inputs> = (data: Inputs) => {
     const updatedValues = updateValues(defaultValues, data);
     const body = convertingValues(updatedValues);
     const dataForLocal = convertingValues({ ...data, ...body });
