+++ src/components/Form/utils/transformData.spec.ts
@@ -5,6 +5,7 @@ describe("transforming values", () => {
   it("should", () => {
     const result = {
       ...data,
+      JWT_SECRET_TTL: 86400,
       EMAIL_PORT: 1025,
       REFRESH_SECRET_TTL: 7,
       OAUTH_GOOGLE_ENABLED: true,
