+++ src/components/Form/validation.ts
@@ -1,4 +1,19 @@
 import { z } from "zod";
+import { TBackendData } from "@components";
+
+export type Inputs = {
+  JWT_SECRET_TTL: number;
+  REFRESH_SECRET_TTL: number;
+  EMAIL_PORT: number;
+  OAUTH_GOOGLE_ENABLED: boolean;
+  LDAP_ENABLED: boolean;
+} & Pick<
+  TBackendData,
+  | "OAUTH_GOOGLE_CLIENT_SECRET"
+  | "OAUTH_GOOGLE_CLIENT_ID"
+  | "LDAP_URL"
+  | "LDAP_BIND_DN"
+>;
 
 export const schema = z.object({
   APP_URL: z.string().url(),
@@ -6,7 +21,7 @@ export const schema = z.object({
   OAUTH_GOOGLE_CLIENT_ID: z.string(),
   OAUTH_GOOGLE_CLIENT_SECRET: z.string(),
   JWT_SECRET: z.string(),
-  JWT_SECRET_TTL: z.string(),
+  JWT_SECRET_TTL: z.number(),
   REFRESH_SECRET: z.string(),
   REFRESH_SECRET_TTL: z
     .number()
