+++ src/components/inputs/Checkbox/Checkbox.tsx
@@ -1,8 +1,7 @@
-"use client";
-import React, { forwardRef, Ref } from "react";
+import { forwardRef, Ref } from "react";
 import { FormControlLabel } from "@mui/material";
 
-import { Props } from "./types";
+import type { Props } from "./types";
 import * as S from "./styled";
 
 const CheckboxComponent = (
