+++ src/components/inputs/NumberInput/NumberInput.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
-import { NumberInput } from "./NumberInput";
+import { NumberInput } from "@components";
 
 const meta: Meta<typeof NumberInput> = {
   title: "Inputs/NumberInput",
