+++ src/components/inputs/TextField/TextField.tsx
@@ -1,7 +1,7 @@
 import { forwardRef, Ref } from "react";
 import { FormControl } from "@mui/material";
 
-import { Props } from "./types";
+import type { Props } from "./types";
 import * as S from "./styled";
 
 const TextFieldComponent = (
