+++ src/components/inputs/index.ts
@@ -1,4 +1,4 @@
+export { Button } from "./Button";
 export { Checkbox } from "./Checkbox";
 export { TextField } from "./TextField";
 export { NumberInput } from "./NumberInput";
-export { Button } from "./Button";
