+++ src/mocks/mockApiHandlers.ts
@@ -1,5 +1,5 @@
 import { rest } from "msw";
-import { data } from "@components/Form/data";
+import { data } from "@components";
 
 export const mockApiHandlers = [
   rest.get("/config", (req, res, ctx) => {
